Refactor: Split dynamic form
- Extract field renderers and row layout from DynamicForm

- Move submit success and duplicate-shop dialogs into focused components

- Keep form submission orchestration in the main island
